tests/e2e/backend-socket-e2e/schemathesis-tests/requirements.txt
schemathesis==3.39.7
python-socketio[client]>=5.10.0
requests>=2.31.0
pytest>=7.4.0
